requirements.txt
httpx
fastmcp
requests>=2.0.0
PyYAML>=6.0
uvicorn>=0.30.0

# Testing dependencies
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
